Transform Dropdown from stylecomponent to normal Component
