+++ web/src/components/Dropdown/Dropdown.module.scss
@@ -1,0 +1,76 @@
+@import "../../styles/mixins.scss";
+
+$spacingOuter: 2rem;
+$spacingInner: 0.5rem;
+$transitionDuration: 0.2s;
+$width: 200px;
+$borderRadius: 2px;
+
+.dropdown {
+  position: relative;
+  width: $width;
+  height: 48px;
+  cursor: pointer;
+
+  .current {
+    display: flex;
+    align-items: center;
+    z-index: 1;
+    padding: 0 $spacingOuter;
+    border: 1px solid var(--colorSeparator);
+    border-radius: 24px;
+    width: 100%;
+    height: 100%;
+
+    @include hoverOverlay();
+
+    .label {
+      width: 100%;
+      white-space: nowrap;
+      text-overflow: ellipsis;
+      overflow: hidden;
+      color: inherit;
+      margin: 0;
+    }
+  }
+
+  .optionsWrapper {
+    position: absolute;
+    right: 0;
+    visibility: hidden;
+    top: calc(100% + 1rem);
+    min-width: $width;
+    border-top: 0;
+    background: var(--colorElevationHigh);
+    box-shadow: var(--dropShadowElevationHigh);
+    opacity: 0;
+    border-radius: $borderRadius;
+    transition: opacity $transitionDuration, visibility 0s $transitionDuration;
+    z-index: 99;
+
+    .triangle {
+      width: 1rem;
+      height: 1rem;
+      background: var(--colorElevationHigh);
+      position: absolute;
+      top: -0.5rem;
+      right: $spacingOuter;
+      transform: rotate(45deg);
+    }
+
+    .options {
+      overflow: hidden;
+      border-radius: $borderRadius;
+    }
+  }
+
+  &.hasIcon .current .label {
+    margin: 0 $spacingOuter 0 0;
+  }
+
+  &.expanded .optionsWrapper {
+    visibility: visible;
+    opacity: 1;
+    transition: opacity $transitionDuration, visibility 0s;
+  }
+}
